move makeVariance
